Avoid not portable %hh convertion specifier. Make idmef-message-print print GMT offset.
git-svn-id: file:///home/yoann/SVN/trunk/libprelude@6324 09c5ec92-17d4-0310-903a-819935f44dba

+++ src/idmef-data.c
@@ -456,7 +456,10 @@ int idmef_data_to_string(const idmef_data_t *data, prelude_string_t *out)
 		break;
 
 	case IDMEF_DATA_TYPE_BYTE:
-		ret = prelude_string_sprintf(out, "%hhu", data->data.byte_data);
+                /*
+                 * %hh convertion specifier is not portable.
+                 */
+		ret = prelude_string_sprintf(out, "%u", (unsigned int) data->data.byte_data);
 		break;
 
 	case IDMEF_DATA_TYPE_UINT32:
